refactor: update imports to use shadcn directory in inspections pages
